oh nO

+++ src/main/java/GUI/GUI_Account.java
@@ -192,11 +192,14 @@ public class GUI_Account extends JPanel {
         });
 
         deleteButton.addActionListener(e -> {
-            if (helped.confirmDelete("Bạn có chắc muốn xóa tài khoản này?")) {
-                AccountDAO.deleteAccount(accountChoosing.getUsername());
-                loadAccounts();
+            if (a.getUsername().equals(accountChoosing.getUsername())) {
+                JOptionPane.showConfirmDialog(null, "Tài khoản này đang được sử dụng vui lòng chọn tài khoản!");
+            } else {
+                if (helped.confirmDelete("Bạn có chắc muốn xóa tài khoản này?")) {
+                    AccountDAO.deleteAccount(accountChoosing.getUsername());
+                    loadAccounts();
+                }
             }
-
         });
 
         searchField.setSearchListener(e -> {
